Skip SWF view creation if the document cannot be converted to PDF (e.g. if the document conversion service is disabled)

+++ src/main/java/org/jahia/modules/docviewer/DocumentViewService.java
@@ -159,11 +159,15 @@ public class DocumentViewService {
      */
     public File convert(File sourceFile, String sourceContentType) throws Exception {
         File pdfFile = convertToPDF(sourceFile, sourceContentType);
-        try {
-            return pdf2swfConverterService.convert(pdfFile);
-        } finally {
-            FileUtils.deleteQuietly(pdfFile);
+        if (pdfFile != null) {
+            try {
+                return pdf2swfConverterService.convert(pdfFile);
+            } finally {
+                FileUtils.deleteQuietly(pdfFile);
+            }
         }
+        
+        return null;
     }
 
     /**
